rename sample_destination json file
